typing for process and split git file
